Fix scripts
