+++ .github/workflows/ci.yml
@@ -39,7 +39,7 @@ jobs:
         run: bun install --frozen-lockfile
 
       - name: Check Formatting [API Server]
-        working-directory: ./packages/api
+        working-directory: ./packages/server
         run: bun run lint
 
       # - name: Check Formatting [Client]
@@ -47,7 +47,7 @@ jobs:
       #   run: TODO: add linting
 
       - name: Run API Tests
-        working-directory: ./packages/api
+        working-directory: ./packages/server
         run: bun run test
 
       - name: Upload coverage reports to Codecov
