+++ packages/server/src/types/server.ts
@@ -1,0 +1,7 @@
+export enum ServerMessages {
+  SERVER_READY = 'SERVER_READY',
+  SERVER_INIT = 'SERVER_INIT',
+  SERVER_ERROR = 'SERVER_ERROR',
+  SERVER_CLOSE = 'SERVER_CLOSE',
+  SERVER_CLOSED = 'SERVER_CLOSED',
+}
